move fixture to conftest.py
